+++ src/ordenes/modulos/aplicacion/comandos/crear_orden.py
@@ -1,29 +1,35 @@
+from abc import ABC, abstractmethod
 from typing import List
 
-from ....seedwork.aplicacion.comandos import Comando, ComandoHandler
-from ....seedwork.aplicacion.comandos import ejecutar_commando as comando
+from ...infraestructura.despachadores import Despachador
 from ....modulos.dominio.entidades import Orden
 from ....modulos.dominio.objetos_valor import Email, Nombre, Producto
+from ....modulos.infraestructura.v1.comandos import ComandoCrearOrden, CrearOrden
 from dataclasses import dataclass
 import datetime
 import time
 
+from ....seedwork.infraestructura import utils
 
-@dataclass
-class ComandoCrearOrden(Comando):
-    nombres: str
-    apellidos: str
-    email: str
-    productos: List[str]
+
+class Comando:
+    ...
+
+
+class ComandoHandler(ABC):
+    @abstractmethod
+    def handle(self, comando: Comando):
+        raise NotImplementedError()
 
 
 class CrearOrdenHandler(ComandoHandler):
 
     def a_entidad(self, comando: ComandoCrearOrden) -> Orden:
         params = dict(
-            nombre=Nombre(comando.nombres, comando.apellidos),
-            email=Email(comando.email),
-            productos=[Producto(nombre, 1) for nombre in comando.productos],
+            nombre_cliente=Nombre(comando.data.nombres, comando.data.apellidos),
+            email=Email(comando.data.email),
+            direccion=comando.data.direccion,
+            productos=[Producto(nombre, 1) for nombre in comando.data.productos],
             fecha_creacion=datetime.datetime.now(),
             fecha_actualizacion=datetime.datetime.now()
         )
@@ -32,9 +38,11 @@ class CrearOrdenHandler(ComandoHandler):
 
     def handle(self, comando: ComandoCrearOrden):
         orden = self.a_entidad(comando)
-
-
-@comando.register(ComandoCrearOrden)
-def ejecutar_comando_crear_orden(comando: ComandoCrearOrden):
-    handler = CrearOrdenHandler()
-    handler.handle(comando)
+        comando = ComandoCrearOrden(
+            time=utils.time_millis(),
+            ingestion=utils.time_millis(),
+            datacontenttype=CrearOrden.__name__,
+            data=orden
+        )
+        despachador = Despachador()
+        despachador.publicar_mensaje(comando, "evento-orden-creada")
