+++ src/ordenes/modulos/dominio/entidades.py
@@ -7,14 +7,14 @@ En este archivo usted encontrará las entidades del dominio de ordenes
 from datetime import datetime
 from typing import List
 
-from ...seedwork.dominio.entidades import Entidad, AgregacionRaiz
+from ...seedwork.dominio.entidades import AgregacionRaiz
 from dataclasses import dataclass, field
 
-from .objetos_valor import Nombre, Email, Direccion, Producto
+from .objetos_valor import Nombre, Email, Producto
 
 @dataclass
 class Orden(AgregacionRaiz):
     nombre_cliente: Nombre = field(default_factory=Nombre)
     email: Email = field(default_factory=Email)
-    direccion: Direccion = field(default_factory=Direccion)
+    direccion: str = field(default_factory=str)
     productos: List[Producto] = field(default_factory=List[Producto])
